// widgets/src/protos/widgets.ts
export interface MiniUserInfo {
  id: string;
  firstName: string;
  lastName: string;
  login: string;
  avatar: string;
}

export interface GetMostLikedUserInfoByUserIdRequest {
  currentUserId: string;
}

export interface GetMostLikedUserInfoByUserIdResponse {
  userInfo: MiniUserInfo | undefined;
}

export interface TagInfo {
  id: string;
  title: string;
  count: number;
}

export interface GetMostLikedTagByUserIdRequest {
  currentUserId: string;
}

export interface GetMostLikedTagByUserIdResponse {
  tagInfo: TagInfo | undefined;
}

export interface GetMostViewedUserInfoByUserIdRequest {
  currentUserId: string;
}

export interface GetMostViewedUserInfoByUserIdResponse {
  userInfo: MiniUserInfo | undefined;
}
// ...
export const GetMostLikedUserInfoByUserIdRequest = {
  encode(message: GetMostLikedUserInfoByUserIdRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.currentUserId !== "") {
      writer.uint32(10).string(message.currentUserId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetMostLikedUserInfoByUserIdRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetMostLikedUserInfoByUserIdRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.currentUserId = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetMostLikedUserInfoByUserIdRequest {
    return { currentUserId: isSet(object.currentUserId) ? globalThis.String(object.currentUserId) : "" };
  },

  toJSON(message: GetMostLikedUserInfoByUserIdRequest): unknown {
    const obj: any = {};
    if (message.currentUserId !== "") {
      obj.currentUserId = message.currentUserId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetMostLikedUserInfoByUserIdRequest>, I>>(
    base?: I,
  ): GetMostLikedUserInfoByUserIdRequest {
    return GetMostLikedUserInfoByUserIdRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetMostLikedUserInfoByUserIdRequest>, I>>(
    object: I,
  ): GetMostLikedUserInfoByUserIdRequest {
    const message = createBaseGetMostLikedUserInfoByUserIdRequest();
    message.currentUserId = object.currentUserId ?? "";
    return message;
  },
};

function createBaseGetMostLikedUserInfoByUserIdResponse(): GetMostLikedUserInfoByUserIdResponse {
  return { userInfo: undefined };
}

export const GetMostLikedUserInfoByUserIdResponse = {
  encode(message: GetMostLikedUserInfoByUserIdResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.userInfo !== undefined) {
      MiniUserInfo.encode(message.userInfo, writer.uint32(10).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetMostLikedUserInfoByUserIdResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetMostLikedUserInfoByUserIdResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.userInfo = MiniUserInfo.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetMostLikedUserInfoByUserIdResponse {
    return { userInfo: isSet(object.userInfo) ? MiniUserInfo.fromJSON(object.userInfo) : undefined };
  },

  toJSON(message: GetMostLikedUserInfoByUserIdResponse): unknown {
    const obj: any = {};
    if (message.userInfo !== undefined) {
      obj.userInfo = MiniUserInfo.toJSON(message.userInfo);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetMostLikedUserInfoByUserIdResponse>, I>>(
    base?: I,
  ): GetMostLikedUserInfoByUserIdResponse {
    return GetMostLikedUserInfoByUserIdResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetMostLikedUserInfoByUserIdResponse>, I>>(
    object: I,
  ): GetMostLikedUserInfoByUserIdResponse {
    const message = createBaseGetMostLikedUserInfoByUserIdResponse();
    message.userInfo = (object.userInfo !== undefined && object.userInfo !== null)
      ? MiniUserInfo.fromPartial(object.userInfo)
      : undefined;
    return message;
  },
};
// ...
export const GetMostLikedTagByUserIdRequest = {
  encode(message: GetMostLikedTagByUserIdRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.currentUserId !== "") {
      writer.uint32(10).string(message.currentUserId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetMostLikedTagByUserIdRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetMostLikedTagByUserIdRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.currentUserId = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetMostLikedTagByUserIdRequest {
    return { currentUserId: isSet(object.currentUserId) ? globalThis.String(object.currentUserId) : "" };
  },

  toJSON(message: GetMostLikedTagByUserIdRequest): unknown {
    const obj: any = {};
    if (message.currentUserId !== "") {
      obj.currentUserId = message.currentUserId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetMostLikedTagByUserIdRequest>, I>>(base?: I): GetMostLikedTagByUserIdRequest {
    return GetMostLikedTagByUserIdRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetMostLikedTagByUserIdRequest>, I>>(
    object: I,
  ): GetMostLikedTagByUserIdRequest {
    const message = createBaseGetMostLikedTagByUserIdRequest();
    message.currentUserId = object.currentUserId ?? "";
    return message;
  },
};

function createBaseGetMostLikedTagByUserIdResponse(): GetMostLikedTagByUserIdResponse {
  return { tagInfo: undefined };
}

export const GetMostLikedTagByUserIdResponse = {
  encode(message: GetMostLikedTagByUserIdResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.tagInfo !== undefined) {
      TagInfo.encode(message.tagInfo, writer.uint32(10).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetMostLikedTagByUserIdResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetMostLikedTagByUserIdResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.tagInfo = TagInfo.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetMostLikedTagByUserIdResponse {
    return { tagInfo: isSet(object.tagInfo) ? TagInfo.fromJSON(object.tagInfo) : undefined };
  },

  toJSON(message: GetMostLikedTagByUserIdResponse): unknown {
    const obj: any = {};
    if (message.tagInfo !== undefined) {
      obj.tagInfo = TagInfo.toJSON(message.tagInfo);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetMostLikedTagByUserIdResponse>, I>>(base?: I): GetMostLikedTagByUserIdResponse {
    return GetMostLikedTagByUserIdResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetMostLikedTagByUserIdResponse>, I>>(
    object: I,
  ): GetMostLikedTagByUserIdResponse {
    const message = createBaseGetMostLikedTagByUserIdResponse();
    message.tagInfo = (object.tagInfo !== undefined && object.tagInfo !== null)
      ? TagInfo.fromPartial(object.tagInfo)
      : undefined;
    return message;
  },
};

function createBaseGetMostViewedUserInfoByUserIdRequest(): GetMostViewedUserInfoByUserIdRequest {
  return { currentUserId: "" };
}

export const GetMostViewedUserInfoByUserIdRequest = {
  encode(message: GetMostViewedUserInfoByUserIdRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.currentUserId !== "") {
      writer.uint32(10).string(message.currentUserId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetMostViewedUserInfoByUserIdRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetMostViewedUserInfoByUserIdRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.currentUserId = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetMostViewedUserInfoByUserIdRequest {
    return { currentUserId: isSet(object.currentUserId) ? globalThis.String(object.currentUserId) : "" };
  },

  toJSON(message: GetMostViewedUserInfoByUserIdRequest): unknown {
    const obj: any = {};
    if (message.currentUserId !== "") {
      obj.currentUserId = message.currentUserId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetMostViewedUserInfoByUserIdRequest>, I>>(
    base?: I,
  ): GetMostViewedUserInfoByUserIdRequest {
    return GetMostViewedUserInfoByUserIdRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetMostViewedUserInfoByUserIdRequest>, I>>(
    object: I,
  ): GetMostViewedUserInfoByUserIdRequest {
    const message = createBaseGetMostViewedUserInfoByUserIdRequest();
    message.currentUserId = object.currentUserId ?? "";
    return message;
  },
};
// ...
export interface WidgetsServiceServer extends UntypedServiceImplementation {
  getMostLikedUserInfoByUserId: handleUnaryCall<
    GetMostLikedUserInfoByUserIdRequest,
    GetMostLikedUserInfoByUserIdResponse
  >;
  getMostLikedTagByUserId: handleUnaryCall<GetMostLikedTagByUserIdRequest, GetMostLikedTagByUserIdResponse>;
  getMostViewedUserInfoByUserId: handleUnaryCall<
    GetMostViewedUserInfoByUserIdRequest,
    GetMostViewedUserInfoByUserIdResponse
  >;
}
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };